Added Dana's Second sample code
"I have created a sample vb.net solution that includes a client and
server application.  The server manages connections from multiple
clients and supports pausing and resuming the sending of data to
individual clients.  Once a client connects, the server sends a packet
of information every 5 seconds that is displayed in the client.  Client
connect/disconnects are displayed in the server window.  The server
presents the list of IPv4 addresses on the computer and allows you to
select which on to listen on.  The client allows you to specify a
computer name or IP address to connect to."