Create danaNotesFeb4
Dana came to the North Garage and helped us figure out how to start reading CAN data from the car. After some trial and error, we found out that the BMS needed to be connected in order for the CAN-USB to actually spit out data. Once we were able to get data, we used Putty on Ethan's laptop and got live can data (but very raw data, just rapid gibberish). 

Our next tasks are to learn the NUSolarCANWatcher code (also known as "Telemetry Logger") on the SmallPC, and to begin to use that code to turn the telemetry logger gibberish into meaningful, useful data. We will be storing the data in SQL and the code is mostly written to do that already.

Our next tasks are to be able to use the code properly, and get the data to be stored in a meaningful format. After we achieve that, we will need to move on to getting that data fed to the Sockets code that will then wirelessly send the data to a base station laptop.

-Akash